requirements.txt
pvporcupine
pyaudio
speechrecognition
requests
google-generativeai
pydantic
google-genai